Refactor to make the AppNexusConncetor the exported class
