+++ src/openads-appnexus/AppNexusError.js
@@ -1,8 +1,8 @@
-export default class AppNexusErrorException extends Error {
+export default class AppNexusError extends Error {
   constructor ({cause, status, position}) {
     super()
     this.message = `Some error ocurred in appnexus with position id: ${position} `
-    this.name = 'AppNexusErrorException'
+    this.name = 'AppNexusError'
     this.status = status
     this.cause = cause
     this.stack = (new Error()).stack
