+++ src/openads-appnexus/AstWrapper.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 
-export default class AppNexusClient {
+export default class AstWrapper {
   static build () {
     require('@schibstedspain/ast')
     return apntag
